Controllers and Reposotories

+++ MovieAPI/src/main/java/com/encora/movieapi/repositories/UserRepository.java
@@ -6,5 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.encora.movieapi.Entities.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+
 }
